+++ core/src/main/java/kafka/server/builders/ReplicaManagerBuilder.java
@@ -33,7 +33,6 @@ import org.apache.kafka.storage.internals.log.LogDirFailureChannel;
 import org.apache.kafka.storage.log.metrics.BrokerTopicStats;
 
 import java.util.Map;
-import java.util.concurrent.atomic.AtomicBoolean;
 
 import scala.Option;
 
@@ -121,7 +120,6 @@ public class ReplicaManagerBuilder {
                              logDirFailureChannel,
                              alterPartitionManager,
                              brokerTopicStats,
-                             new AtomicBoolean(false),
                              Option.empty(),
                              Option.empty(),
                              Option.empty(),
