+++ core/src/test/scala/unit/kafka/cluster/PartitionLockTest.scala
@@ -27,7 +27,8 @@ import kafka.server._
 import kafka.utils._
 import org.apache.kafka.common.config.TopicConfig
 import org.apache.kafka.common.record.{MemoryRecords, SimpleRecord}
-import org.apache.kafka.common.requests.{FetchRequest, LeaderAndIsrRequest}
+import org.apache.kafka.common.requests.FetchRequest
+import org.apache.kafka.common.PartitionState
 import org.apache.kafka.common.utils.Utils
 import org.apache.kafka.common.{TopicPartition, Uuid}
 import org.apache.kafka.coordinator.transaction.TransactionLogConfig
@@ -142,8 +143,7 @@ class PartitionLockTest extends Logging {
     val replicaToCheck = 3
     val firstReplicaSet = util.List.of[Integer](3, 4, 5)
     val secondReplicaSet = util.List.of[Integer](1, 2, 3)
-    def partitionState(replicas: util.List[Integer]) = new LeaderAndIsrRequest.PartitionState()
-      .setControllerEpoch(1)
+    def partitionState(replicas: util.List[Integer]) = new PartitionState()
       .setLeader(replicas.get(0))
       .setLeaderEpoch(1)
       .setIsr(replicas)
@@ -344,13 +344,11 @@ class PartitionLockTest extends Logging {
 
     partition.createLogIfNotExists(isNew = false, isFutureReplica = false, offsetCheckpoints, Some(topicId))
 
-    val controllerEpoch = 0
     val replicas = (0 to numReplicaFetchers).map(i => Integer.valueOf(brokerId + i)).toList.asJava
     val isr = replicas
     replicas.forEach(replicaId => when(metadataCache.getAliveBrokerEpoch(replicaId)).thenReturn(Optional.of(1L)))
 
-    assertTrue(partition.makeLeader(new LeaderAndIsrRequest.PartitionState()
-      .setControllerEpoch(controllerEpoch)
+    assertTrue(partition.makeLeader(new PartitionState()
       .setLeader(brokerId)
       .setLeaderEpoch(leaderEpoch)
       .setIsr(isr)
